refactor validator to be a base class with get_version function defined
