+++ cdisc_rules_engine/interfaces/terms_factory_interface.py
@@ -1,8 +1,6 @@
 from abc import ABC, abstractmethod
-from typing import Dict
 
 from cdisc_rules_engine.interfaces import DataServiceInterface
-from cdisc_rules_engine.interfaces import DictionaryTermInterface
 
 
 class TermsFactoryInterface(ABC):
@@ -20,9 +18,13 @@ class TermsFactoryInterface(ABC):
     def install_terms(
         self,
         directory_path: str,
-    ) -> Dict[
-        str, Dict[str, DictionaryTermInterface]
-    ]:  # maps term type to a dictionary of term identifiers to term
+    ):  # maps term type to a dictionary of term identifiers to term
         """
         Accepts file contents and saves it to the DB.
+        """
+
+    @abstractmethod
+    def get_version(self, directory_path: str) -> str:
+        """
+        Reads version from the provided dictionary files
         """
